requirements.txt
numpy
Pillow==9.0.0
